refactor: api 데이터 타입 분리
#33
