+++ src/containers/study/Participant/types.ts
@@ -1,10 +1,4 @@
-export interface ParticipantType {
-  id: number;
-  name: string;
-  status: string;
-  profileImg: string;
-  myPageUrl: string;
-}
+import { ParticipantType } from '@/types';
 
 export interface ParticipantProps {
   participantInfos: ParticipantType[];
