+++ src/types.ts
@@ -10,3 +10,11 @@ export interface GardenInfoType {
   count: number;
   id: number;
 }
+
+export interface ParticipantType {
+  id: number;
+  name: string;
+  status: string;
+  profileImg: string;
+  myPageUrl: string;
+}
